mcp_sample/sample_c/requirements.txt
fastmcp>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
mcp>=1.0.0 